+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.26...gatsby-plugin-typescript@2.1.27) (2020-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.1.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.25...gatsby-plugin-typescript@2.1.26) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.1.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.23...gatsby-plugin-typescript@2.1.25) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.1.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.23...gatsby-plugin-typescript@2.1.24) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.1.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.22...gatsby-plugin-typescript@2.1.23) (2019-12-12)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.1.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.20...gatsby-plugin-typescript@2.1.22) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.1.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.20...gatsby-plugin-typescript@2.1.21) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
 ## [2.1.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.19...gatsby-plugin-typescript@2.1.20) (2019-11-26)
 
 **Note:** Version bump only for package gatsby-plugin-typescript